+++ Res/Klase/PunjacAutomobila.cs
@@ -10,9 +10,9 @@ namespace Klase
     {
         public string Naziv { get; set; }
 
-        public int TrenutnoBaterijaAutomobila { get; set; }//procenat punosti
-        public int SnagaBaterijePunjaca { get; set; } //koliko moze da puni punjac po satu
-        public int MaksBaterijaAutomobila { get; set; }// dokle se puni
+        public double TrenutnoBaterijaAutomobila { get; set; }//procenat punosti
+        public double SnagaBaterijePunjaca { get; set; } //koliko moze da puni punjac po satu
+        public double MaksBaterijaAutomobila { get; set; }// dokle se puni
 
         public bool UtaknutAutomobil { get; set; } = false;
 
@@ -21,7 +21,7 @@ namespace Klase
 
         }
 
-        public PunjacAutomobila(string naziv, int snagaBaterije, int procenatBaterije, int maksSnagabaterijeAuta)
+        public PunjacAutomobila(string naziv, double snagaBaterije, double procenatBaterije, double maksSnagabaterijeAuta)
         {
             if (String.IsNullOrEmpty(naziv))
             {
